Articles Table now displaying
